refactor(Invoice.swift, PDF.swift, Product.swift, ProductDetails.swift, StandaloneInvoice.swift, StandaloneItem.swift, StandaloneItemDetails.swift): change price type from Int to Float to support decimal values in product prices feat(PDF.swift): add formattedNumber function overload for Float input to support formatted output of decimal prices fix(Invoice.swift, StandaloneInvoice.swift): adjust total calculation to use Float multiplication for quantity and price fix(PDF.swift): adjust table content generation to use Float multiplication for quantity and price

+++ Invoice Maker/Model/Product/Product.swift
@@ -13,10 +13,10 @@ class Product {
     @Attribute(.unique) var code: Int
     var name: String
     var details: String?
-    var price: Int
+    var price: Float
     var createdDate: Date = Date.now
 
-    init(code: Int, name: String, details: String?, price: Int) {
+    init(code: Int, name: String, details: String?, price: Float) {
         self.code = code
         self.name = name
         self.details = details
