+++ Invoice Maker/Model/StandaloneInvoice/StandaloneInvoice.swift
@@ -33,8 +33,8 @@ class StandaloneInvoice {
         number.isEmpty || customerId == nil || items.isEmpty
     }
 
-    var total: Int {
-        items.reduce(0) { $0 + ($1.quantity * $1.productPrice) }
+    var total: Float {
+        items.reduce(0) { $0 + (Float($1.quantity) * $1.productPrice) }
     }
 
     init(number: String = "",
